feat: Add verbose option to transcribe_audio_to_text_with_timestamps function
The code changes add a new optional `verbose` parameter to the `transcribe_audio_to_text_with_timestamps` function. This allows the user to control the verbosity of the function by printing progress messages during the transcription process. The `verbose` parameter is also passed to the `transcribe_audio_to_text_with_timestamps` function in the `process_directory` function, allowing the user to enable or disable verbosity for the entire batch transcription process.